add a legacy app middleware to pipeline
I have a use case where I need to load a legacy app to sit on top of
mezzio front-end, where the legacy app has its own routing structure.

Legacy app is to load when a route is called that is not familiar to
mezzio, and before mezzio issues a 404 error

This middleware acts as a skeleton to use when writing code to load
a legacy app with its own routing structure.

The idea is that mezzio and legacy app will peacefully coexist this way
together.  This relies on mezzio app being able to authenticate
successfully without triggering a 404 error.